chore(compass-e2e-tests): refactor out common favorite operations
