+++ packages/compass-e2e-tests/helpers/commands/select-favorite.ts
@@ -1,0 +1,13 @@
+import type { CompassBrowser } from '../compass-browser';
+import * as Selectors from '../selectors';
+
+export async function selectFavorite(
+  browser: CompassBrowser,
+  favoriteName: string
+): Promise<void> {
+  await browser.clickVisible(Selectors.sidebarFavoriteButton(favoriteName));
+  await browser.waitUntil(async () => {
+    const text = await browser.$(Selectors.ConnectionTitle).getText();
+    return text === favoriteName;
+  });
+}
